# p4est surface performance 3D