--- a/src/solvers/dgsem_p4est/dg.jl
+++ b/src/solvers/dgsem_p4est/dg.jl
@@ -27,23 +27,6 @@ function create_cache(mesh::P4estMesh, equations::AbstractEquations, dg::DG, ::A
 end
 
 
-# TODO: p4est interface performance, remove
-# Extract outward-pointing normal vector (contravariant vector ±Ja^i, i = index) as SVector
-# Note that this vector is not normalized
-@inline function get_normal_vector(direction, cache, indices...)
-  @unpack contravariant_vectors = cache.elements
-
-  orientation = div(direction + 1, 2)
-  normal = get_contravariant_vector(orientation, contravariant_vectors, indices...)
-
-  # Contravariant vectors at interfaces in negative coordinate direction are pointing inwards
-  if direction in (1, 3, 5)
-    normal *= -1
-  end
-
-  return normal
-end
-
 # Extract outward-pointing normal direction
 # (contravariant vector ±Ja^i, i = index)
 # Note that this vector is not normalized