… examples/tree_3d_dgsem/elixir_advection_extended.jl